Menu: refactor MenuItem and MenuGroup to be standalone component
